+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 
 ## [1.1.0](https://github.com/futurestudio/hapi-rate-limitor/compare/v1.0.0...v1.1.0) - 2019-11-xx
 
+### Added
+- checks to ensure a token when decoding a JWT
+- checks to ensure a payload when creating a JWT
+
 ### Updated
 - moved TypeScript definitions file to `typings/index.d.ts`